feat: add dedicated bootstrap page to admin console
