+++ docs/01_project/activeContext/tasks/completed/2026-02-25.md
@@ -96,3 +96,18 @@
 - [x] `NPM_CONFIG_PREFIX=/tmp/npm-global gh act --workflows .github/workflows/test.yml --job format-check`（pass）
 - [x] `NPM_CONFIG_PREFIX=/tmp/npm-global gh act --workflows .github/workflows/test.yml --job native-test-linux`（pass）
 - [x] `NPM_CONFIG_PREFIX=/tmp/npm-global gh act --workflows .github/workflows/test.yml --job community-node-tests`（pass）
+
+## Issue #164: Community Node Admin UI の Bootstrap 専用ページ化
+
+- [x] `admin-console` サイドバー内の Bootstrap 詳細カードを撤去し、専用ページ遷移に変更。
+- [x] サイドバーへ `Bootstrap` ナビゲーション項目を追加し、`/bootstrap` ルートを新設。
+- [x] 専用ページで `node_id@host:port` 形式表示、接続ユーザー一覧、接続ユーザー数表示を実装（0件時の空表示を含む）。
+- [x] `App.test.tsx` を新導線に合わせて更新し、`BootstrapPage.test.tsx` を追加して表示要件回帰を固定化。
+- [x] 進捗レポート `docs/01_project/progressReports/2026-02-25_issue164_admin_bootstrap_page.md` を追加。
+
+## 検証（Issue #164）
+
+- [x] `docker run --rm -v /home/kingyosun/kukuri:/workspace -w /workspace/kukuri-community-node/apps/admin-console node:22-bullseye bash -lc "set -euo pipefail; corepack enable; pnpm --version; pnpm install --frozen-lockfile; pnpm test src/App.test.tsx src/pages/BootstrapPage.test.tsx; pnpm typecheck"`（pass）
+- [x] `NPM_CONFIG_PREFIX=/tmp/npm-global XDG_CACHE_HOME=/tmp/xdg-cache DOCKER_CONFIG=/tmp/docker-config ACT_CACHE_DIR=/tmp/act-cache gh act --workflows .github/workflows/test.yml --job format-check 2>&1 | tee tmp/logs/gh-act-format-check-issue164.log`（pass）
+- [x] `NPM_CONFIG_PREFIX=/tmp/npm-global XDG_CACHE_HOME=/tmp/xdg-cache DOCKER_CONFIG=/tmp/docker-config ACT_CACHE_DIR=/tmp/act-cache gh act --workflows .github/workflows/test.yml --job native-test-linux 2>&1 | tee tmp/logs/gh-act-native-test-linux-issue164.log`（pass）
+- [x] `NPM_CONFIG_PREFIX=/tmp/npm-global XDG_CACHE_HOME=/tmp/xdg-cache DOCKER_CONFIG=/tmp/docker-config ACT_CACHE_DIR=/tmp/act-cache gh act --workflows .github/workflows/test.yml --job community-node-tests 2>&1 | tee tmp/logs/gh-act-community-node-tests-issue164.log`（pass）
